+++ app/views/tops/show.html.haml
@@ -1,3 +1,1 @@
 %h1 Topページ
-= link_to '時間毎予約率', reservation_rate_for_time_path
-= link_to '講師毎予約率', reservation_rate_for_teacher_path
